requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.2.1
et-xmlfile==1.1.0
Flask==3.1.1
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.0
PyMySQL==1.1.1
python-dateutil==2.8.2
python-dotenv==1.1.1
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.1.3
